requirements.txt
beautifulsoup4>=4.9.3
nltk>=3.6.2
python-dotenv>=0.19.0 